+++ docs/history.md
@@ -4,6 +4,8 @@
 
 - Fixed the bug where the HUD had overlapping sprites if you were racing a multi-character speedrun. (Thanks Gamonymous)
 - Fixed the bug where the speedrun timer would overlap with the race timer if you were racing a multi-character speedrun. (The race timer will take precedence.) (Thanks Gamonymous)
+- Fixed the bug where a Checkpoint could rotate if you were playing as Tainted Isaac and you re-entered the room with the Checkpoint. (Thanks 91)
+- Fixed the bug where the free devil deal icon would not show properly when resetting the game.
 
 ### *v0.77.1* - March 6th, 2022